Move csv defs from `data_definitions` to `inputs` dir
